+++ lib/contentful/management/client_asset_methods_factory.rb
@@ -1,5 +1,4 @@
 require_relative 'client_association_methods_factory'
-require_relative 'client_association_all_published_method_factory'
 
 module Contentful
   module Management
@@ -7,7 +6,6 @@ module Contentful
     # @private
     class ClientAssetMethodsFactory
       include Contentful::Management::ClientAssociationMethodsFactory
-      include Contentful::Management::ClientAssociationAllPublishedMethodsFactory
     end
   end
 end